// src/scripts/types/interfaces.ts
export interface Project {
    deployUrl: string;
    projectName: string;
    techStack: string[];
    description: string;
    repo: string;
    back?: string;
}

export interface TextFields {
    navElems: string[];
    name: string;
    spec: string;
    languageEn: string;
    languageRu: string;
    aboutText: string;
    momentumDescr: string;
    songBirdDescr: string;
    gemPuzzleDescr: string;
    shopDescr: string;
    seaBattleDescr: string;
    code: string;
    deploy: string;
    back: string;
    educationTextOne: string;
    location: string;
    languageTitle: string;
}